Move ES5-only mapping tests to ElasticSearch5/ subfolder
AbstractMappingTestCase and UpdateEvent/Organizer/PlaceMappingTest moved
to ElasticSearch5/ as these operations are ES5-only. Concrete test classes
implement the ElasticSearch5Test marker interface.

UpdateRegionMappingTest in ElasticSearch5/ now inherits the ES5 test from
AbstractMappingTestCase, removing the previously duplicate explicit test.
